change to restlike urls, more tests, partly error handling.
1: project administration 

Task-Url: http://github.com/man-at-home/totask2/issues/issue/1

+++ src/main/java/org/regele/totask2/controller/TaskController.java
@@ -2,13 +2,14 @@ package org.regele.totask2.controller;
 
 import java.util.List;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.regele.totask2.model.Project;
 import org.regele.totask2.model.ProjectRepository;
 import org.regele.totask2.model.Task;
 import org.regele.totask2.model.TaskRepository;
+import org.regele.totask2.util.ProjectNotFoundException;
+import org.regele.totask2.util.TaskNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,8 +17,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 
 
 /** 
@@ -44,61 +47,60 @@ public class TaskController {
     private ProjectRepository projectRepository;    
 
     /** simple hello world. */
-    @RequestMapping(value = "/tasks", params = {"tasksByProject"}, method = RequestMethod.GET)
-    public String tasks(final Model model, final HttpServletRequest req) {
+    @RequestMapping(value = "/project/{id}/tasks", method = RequestMethod.GET)
+    public String tasksForProject(@PathVariable final long id, final Model model) {
         
-        LOG.debug("tasks");
+        LOG.debug("tasks for project " + id);
    
-        long projectId = Long.parseLong( req.getParameter("tasksByProject") );
-       
-        List<Task> tasks = taskRepository.findByProjectId(projectId);
+        List<Task> tasks = taskRepository.findByProjectId(id);
         
         model.addAttribute("tasks", tasks);           
-        model.addAttribute("projectId", projectId);
+        model.addAttribute("projectId", id);
         
-        LOG.debug("serving " + tasks.size() + " tasks for project " + projectId);
+        LOG.debug("serving " + tasks.size() + " tasks for project " + id);
         return "tasks";
     }
     
     /** show edit page for an existing task. GET. */
-    @RequestMapping(value = "/editTask", params = {"editTask"}, method = RequestMethod.GET)
-    public String editTask(final Model model, final HttpServletRequest req) {
+    @RequestMapping(value = "/task/{id}", method = RequestMethod.GET)
+    public String editTask(@PathVariable final long id, final Model model) {
         
-        LOG.debug("edit task");
+        LOG.debug("edit task " + id);
         
-        long taskId = Long.parseLong( req.getParameter("editTask") );
-        Task task = taskRepository.getOne(taskId);
+        Task task = taskRepository.getOne(id);
         model.addAttribute("task", task);        
         
         LOG.debug("will edit " + task);
         return "editTask";
     }
     
-    /** show edit page for an existing task. GET. */
-    @RequestMapping(value = "/editTask", params = {"deleteTask"}, method = RequestMethod.GET)
-    public String deleteTask(final Model model, final HttpServletRequest req) {
+    /** delete an existing task. POST. */
+    @RequestMapping(value = "task/delete", method = RequestMethod.POST)
+    public String deleteTask(@RequestParam long id, final Model model) {
         
-        LOG.debug("delete task");
+        LOG.debug("delete task " + id);
         
-        long taskId = Long.parseLong( req.getParameter("deleteTask") );
-        Task taskToDelete = taskRepository.getOne(taskId);
+        Task taskToDelete = taskRepository.findOne(id);
+        if( taskToDelete == null)
+            throw new TaskNotFoundException("task " + id + " not found for deletion.");
+        
         long projectId = taskToDelete.getProject().getId();
         
         taskRepository.delete(taskToDelete);
         taskRepository.flush();
         
-        LOG.debug("deleted task " + taskId);
-        return "redirect:/tasks?tasksByProject=" + projectId;
+        LOG.debug("deleted task " + id);
+        return "redirect:/project/" + projectId + "/tasks";
     }
     
-    /** show edit page for a new project. GET. */
-    @RequestMapping(value = "/newTask", params = {"projectId"}, method = RequestMethod.GET)
-    public String newTask(final Model model, final HttpServletRequest req) {       
-        LOG.debug("new task");
+    /** show edit page for a new task. GET. */
+    @RequestMapping(value = "project/{projectId}/task/new", method = RequestMethod.GET)
+    public String newTask(@PathVariable final long projectId, final Model model) {       
+        LOG.debug("new task for project " + projectId);
         
-        long projectId = Long.parseLong( req.getParameter("projectId") );
-        Project project = this.projectRepository.getOne(projectId);
-        
+        Project project = this.projectRepository.findOne(projectId);
+        if( project == null)
+            throw new ProjectNotFoundException("project " + projectId + " not found creating new task.");
         Task task = project.createTask();
         model.addAttribute("task", task);        
         
@@ -108,8 +110,8 @@ public class TaskController {
     
     
     /** save edited task. POST. */
-    @RequestMapping(value = "/editTask", method = RequestMethod.POST)
-    public String editTaskSave(@Valid final Task task, final BindingResult bindingResult, final ModelMap model, final HttpServletRequest req) {        
+    @RequestMapping(value = "/task/save", method = RequestMethod.POST)
+    public String editTaskSave(@Valid final Task task, final BindingResult bindingResult, final ModelMap model) {        
         LOG.debug("editTaskSave(" + task + ")");
         
         if (bindingResult.hasErrors()) {
@@ -120,7 +122,7 @@ public class TaskController {
         model.clear();        
         
         LOG.debug("saved " + savedTask + ", now redirecting.");
-        return "redirect:/tasks?tasksByProject=" + savedTask.getProject().getId();
+        return "redirect:/project/" + savedTask.getProject().getId() + "/tasks";
     }    
     
 }
